pyproject.toml
[project]
name = "fairglow"
version = "0.1.0"
description = "Take home project"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "flask-sqlalchemy>=3.1.1",
]
